pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "direct-cli"
dynamic = ["version"]
description = "Command-line interface for Yandex Direct API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.6"
authors = [
    {name = "Pavel Maksimov", email = "vur21@ya.ru"}
]
keywords = ["yandex", "direct", "cli", "api", "advertising"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "tapi-yandex-direct>=2021.5.29",
    "click>=8.0.0",
    "python-dotenv>=0.19.0",
    "tabulate>=0.8.0",
    "colorama>=0.4.0",
    "tqdm>=4.60.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=22.0",
    "flake8>=4.0",
]

[project.scripts]
direct-cli = "direct_cli.cli:cli"

[project.urls]
Homepage = "https://github.com/pavelmaksimov/direct-cli"
Repository = "https://github.com/pavelmaksimov/direct-cli"

[tool.setuptools_scm]
write_to = "direct_cli/_version.py"

[tool.setuptools.packages.find]
where = ["."]
include = ["direct_cli*"]

[tool.black]
line-length = 88
target-version = ['py36']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
